Fixed FV1.21.3 user longitude / latitude storage

+++ init/framework/1.21.3.php
@@ -3,6 +3,6 @@
  * Fix servers table, createdby may be NULL
  */
 sql_query('ALTER TABLE `users`
-           CHANGE COLUMN `latitude`  `latitude`  FLOAT(14, 14) NULL DEFAULT NULL,
-           CHANGE COLUMN `longitude` `longitude` FLOAT(14, 14) NULL DEFAULT NULL');
+           CHANGE COLUMN `latitude`  `latitude`  DECIMAL(18, 15) NULL DEFAULT NULL,
+           CHANGE COLUMN `longitude` `longitude` DECIMAL(18, 15) NULL DEFAULT NULL');
 ?>
